Updated when looking for online data, romc does not perform useless requests.
Now each API have a state, telling whether data were find or not.

If you run romc, and you use JVC API, but there is no data for certains games,
romc will save that state. So if you re-run the script with JVC, romc wont
look for data on those games.

But if you use a differente API, this time romc will look for data.

This is reduce the number of requests.